Hash custom_data in state storage
This also switches to helpers for b64.
